refactor GCE and EC2 checks
Check the facter version, if 3.0+ then utilise the built in facts for
ec2 (the ec2 fact format changed in v3) and gce. Else try and establish
the facts via the old facter method for ec2, or from the ice metadata
api.

Also cast the providerId to a string to ensure the lookup completes
successfully when querying the API.